fix(codex): keep quota bar after usage limit is hit
After Codex hits the 5-hour cap it writes a later token_count with
primary/secondary JSON null (limit_id flips to "premium"). The parser
took that event as latest, so the row flashed a full red bar then
collapsed to a gray "no limit" bar. Keep the last account-level
windows across the scan window and decode used_percent as a float so
live 98.0 values still parse.

+++ AgentsUsageBar/Providers/Codex/Models/CodexUsageResponse.swift
@@ -68,8 +68,9 @@ public struct CodexUsageResponse: Decodable, Sendable, Equatable {
     /// `limit_window_seconds` instead of `window_minutes`.
     public struct Window: Decodable, Sendable, Equatable {
 
-        /// 0–100 percentage. Caller normalises to a 0.0–1.0 fraction for the UI.
-        public let usedPercent: Int?
+        /// 0–100 percentage. JSON may be an int (`48`) or a float (`48.0`).
+        /// Caller normalises to a 0.0–1.0 fraction for the UI.
+        public let usedPercent: Double?
 
         /// Absolute Unix epoch **seconds** at which this window resets.
         public let resetAt: Int?
